feat: add event info error report button
