+++ src/components/feature/bugReport/BugReportModal.module.css
@@ -1,0 +1,40 @@
+.overlay {
+	position: fixed;
+	inset: 0;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	padding: 24px;
+	background-color: rgba(0, 0, 0, 0.35);
+	z-index: 1400;
+}
+
+.modal {
+	position: relative;
+	width: min(100%, 480px);
+	max-height: calc(100vh - 48px);
+	overflow-y: auto;
+	border-radius: 14px;
+	background-color: #ffffff;
+	box-shadow: 0 12px 32px rgba(0, 0, 0, 0.2);
+}
+
+.closeButton {
+	position: absolute;
+	top: 12px;
+	right: 12px;
+	display: inline-flex;
+	align-items: center;
+	justify-content: center;
+	width: 32px;
+	height: 32px;
+	padding: 0;
+	border: none;
+	border-radius: 8px;
+	background: transparent;
+	cursor: pointer;
+}
+
+.closeButton:hover {
+	background-color: #f5f5f5;
+}
